requirements.txt
certifi==2022.12.7
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
fonttools==4.25.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
mpyq==0.2.5
munkres==1.1.4
numpy==1.24.2
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pip==22.3.1
ply==3.11
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
sc2reader==1.8.0
scikit-learn==1.2.1
scipy==1.10.1
setuptools==65.6.3
sip==6.6.2
six==1.16.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
wheel==0.38.4
wincertstore==0.2
